moved File-related-methods to FilePermissionModule
