+++ src/main/java/org/intellimate/izou/security/SecurityManager.java
@@ -131,95 +131,6 @@ public final class SecurityManager extends java.lang.SecurityManager {
     }
 
     /**
-     * Determines if the file at the given file path is safe to read from in all aspects, if so returns true, else false
-     *
-     * @param filePath the path to the file to read from
-     */
-    private void fileReadCheck(String filePath) {
-        File potentialFile = new File(filePath);
-        String canonicalPath;
-        try {
-            canonicalPath = potentialFile.getCanonicalPath();
-        } catch (IOException e) {
-            logger.error("Error getting canonical path", e);
-            throw getException(filePath);
-        }
-
-        for (String file : allowedReadFiles) {
-            if (canonicalPath.contains(file)) {
-                return;
-            }
-        }
-
-        boolean allowedDirectory = false;
-        for (String dir : allowedReadDirectories) {
-            if (canonicalPath.contains(dir)) {
-                allowedDirectory = true;
-                break;
-            }
-        }
-        if (!allowedDirectory) {
-            throw getException(filePath);
-        }
-
-        String[] pathParts = canonicalPath.split(File.separator);
-        String lastPathPart = pathParts[pathParts.length - 1].toLowerCase();
-
-        String[] pathPeriodParts = lastPathPart.split("\\.");
-        String fileExtension = pathPeriodParts[pathPeriodParts.length - 1].toLowerCase();
-
-        if (!secureAccess.checkForExistingFileOrDirectory(canonicalPath)
-                || secureAccess.checkForDirectory(canonicalPath)) {
-            return;
-        }
-
-        Pattern pattern = Pattern.compile(allowedReadFileTypesRegex);
-        Matcher matcher = pattern.matcher(fileExtension);
-        if (!matcher.matches() || fileExtension.equals(lastPathPart))
-            throw getException(filePath);
-    }
-
-    /**
-     * Determines if the file at the given file path is safe to write to in all aspects, if so returns true, else false
-     *
-     * @param filePath the path to the file to write to
-     */
-    private void fileWriteCheck(String filePath) {
-        File potentialFile = new File(filePath);
-        String canonicalPath;
-        try {
-            canonicalPath = potentialFile.getCanonicalPath();
-        } catch (IOException e) {
-            logger.error("Error getting canonical path", e);
-            throw getException(filePath);
-        }
-
-        boolean allowedDirectory = false;
-        for (String dir : allowedWriteDirectories) {
-            if (canonicalPath.contains(dir)) {
-                allowedDirectory = true;
-                break;
-            }
-        }
-        if (!allowedDirectory) {
-            throw getException(filePath);
-        }
-
-        String[] pathParts = canonicalPath.split("\\.");
-        String fileExtension = pathParts[pathParts.length - 1].toLowerCase();
-
-        if (!secureAccess.checkForExistingFileOrDirectory(canonicalPath)
-                || secureAccess.checkForDirectory(canonicalPath)) {
-            return;
-        }
-
-        Pattern pattern = Pattern.compile(allowedWriteFileTypesRegex);
-        Matcher matcher = pattern.matcher(fileExtension);
-        if (!matcher.matches())
-            throw getException(filePath);
-    }
-
-    /**
      * Checks if {@link SecureAccess} is included in the current class context, if so true is returned, else false
      *
      * @return true if {@link SecureAccess} is included in the current class context, else false
